+++ lib/kuby/utils.rb
@@ -1,5 +1,6 @@
 module Kuby
   module Utils
-    autoload :Table, 'kuby/utils/table'
+    autoload :SemVer, 'kuby/utils/sem_ver'
+    autoload :Table,  'kuby/utils/table'
   end
 end